requirements.txt
Jinja2
weasyprint
jsonschema
python-dotenv
tqdm
